Allow children prop on connected class components
The children props is automatically injected on class components by the
React typings but is required to be explicitly specified for functional
components. Since react-redux switched the type of the connect wrapper
to be a functional component you can no longer pass children to
connected components unless the wrapped component explicitly declares
children on its props. This represents a breaking change in behavior for
class components since they are expected to be able to automatically
take children.

This change explicitly adds the children prop to the connected component
when the wrapped component is a class component.